# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelGPT API",
    "description": "SentinelGPT: AI-Powered Vulnerability Analysis Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/targets": {
      "post": {
        "tags": [
          "Targets"
        ],
        "summary": "Register a new target",
        "description": "Register a target for the requesting user.",
        "operationId": "create_target_api_v1_targets_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTargetRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Targets"
        ],
        "summary": "List targets owned by the current user",
        "operationId": "list_targets_api_v1_targets_get",
        "parameters": [
          {
            "name": "includeArchived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Includearchived"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/targets/{target_id}": {
      "get": {
        "tags": [
          "Targets"
        ],
        "summary": "Get target detail",
        "operationId": "get_target_api_v1_targets__target_id__get",
        "parameters": [
          {
            "name": "target_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Target Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Targets"
        ],
        "summary": "Update target metadata (hostname/URL immutable)",
        "operationId": "update_target_api_v1_targets__target_id__patch",
        "parameters": [
          {
            "name": "target_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Target Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTargetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Targets"
        ],
        "summary": "Archive (soft-delete) a target",
        "description": "Sets isArchived=true. Hard delete is never permitted once scan lifecycle data references the target (SRS Chapter 4, Section 13).",
        "operationId": "delete_target_api_v1_targets__target_id__delete",
        "parameters": [
          {
            "name": "target_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Target Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/dashboard/targets": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Per-target posture",
        "operationId": "get_targets_posture_api_v1_dashboard_targets_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetsPostureResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTargetRequest": {
        "properties": {
          "hostname": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Hostname"
          },
          "url": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "hostname",
          "url"
        ],
        "title": "CreateTargetRequest",
        "description": "POST /targets request body (SRS Chapter 5, Section 4)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LatestScanDto": {
        "properties": {
          "scanId": {
            "type": "string",
            "title": "Scanid"
          },
          "targetId": {
            "type": "string",
            "title": "Targetid"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "completedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          }
        },
        "type": "object",
        "required": [
          "scanId",
          "targetId",
          "status"
        ],
        "title": "LatestScanDto"
      },
...
      "PageInfo": {
        "properties": {
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          },
          "hasNextPage": {
            "type": "boolean",
            "title": "Hasnextpage"
          }
        },
        "type": "object",
        "required": [
          "nextCursor",
          "hasNextPage"
        ],
        "title": "PageInfo",
        "description": "Cursor-pagination metadata (SRS Chapter 5, Section 16)."
      },
...
      "TargetListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TargetResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "pageInfo": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "type": "object",
        "required": [
          "items",
          "pageInfo"
        ],
        "title": "TargetListResponse",
        "description": "Paginated list envelope (SRS Chapter 5, Section 16)."
      },
      "TargetPostureDto": {
        "properties": {
          "targetId": {
            "type": "string",
            "title": "Targetid"
          },
          "hostname": {
            "type": "string",
            "title": "Hostname"
          },
          "isArchived": {
            "type": "boolean",
            "title": "Isarchived"
          },
          "latestScan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LatestScanDto"
              },
              {
                "type": "null"
              }
            ]
          },
          "previousScanId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previousscanid"
          },
          "lastScanAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastscanat"
          },
          "openFindingsTotal": {
            "type": "integer",
            "title": "Openfindingstotal"
          },
          "severityCounts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Severitycounts"
          },
          "priorityCounts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Prioritycounts"
          },
          "regressionsCount": {
            "type": "integer",
            "title": "Regressionscount"
          },
          "remediationCounts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Remediationcounts"
          },
          "doneOpenCount": {
            "type": "integer",
            "title": "Doneopencount"
          }
        },
        "type": "object",
        "required": [
          "targetId",
          "hostname",
          "isArchived",
          "openFindingsTotal",
          "severityCounts",
          "priorityCounts",
          "regressionsCount",
          "remediationCounts",
          "doneOpenCount"
        ],
        "title": "TargetPostureDto"
      },
      "TargetResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "hostname": {
            "type": "string",
            "title": "Hostname"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "ownerUserId": {
            "type": "string",
            "format": "uuid",
            "title": "Owneruserid"
          },
          "isArchived": {
            "type": "boolean",
            "title": "Isarchived"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "PENDING_ATTESTATION"
          }
        },
        "type": "object",
        "required": [
          "id",
          "hostname",
          "url",
          "ownerUserId",
          "isArchived",
          "createdAt"
        ],
        "title": "TargetResponse",
        "description": "Target representation incl. derived attestation status.\n\n``status`` is always PENDING_ATTESTATION until the authorization\nattestation feature lands (Chapter 5 Section 4/5): a target exists but is\nunscannable until a CONFIRMED attestation is presented."
      },
      "TargetsPostureResponse": {
        "properties": {
          "targets": {
            "items": {
              "$ref": "#/components/schemas/TargetPostureDto"
            },
            "type": "array",
            "title": "Targets"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "priorityVersion": {
            "type": "string",
            "title": "Priorityversion"
          }
        },
        "type": "object",
        "required": [
          "targets",
          "total",
          "priorityVersion"
        ],
        "title": "TargetsPostureResponse"
      },
...
      "UpdateTargetRequest": {
        "properties": {
          "isArchived": {
            "type": "boolean",
            "title": "Isarchived"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "isArchived"
        ],
        "title": "UpdateTargetRequest",
        "description": "PATCH /targets/{targetId} body \u2014 hostname/URL are immutable."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}